uint64 to size_t

+++ source/rshash3.cpp
@@ -552,7 +552,7 @@ inline bool RSHash3::check(const uint64_t kmer, const uint64_t kmer_rc,
     uint64_t* offsets, const size_t p, const size_t no_skmers,
     const size_t left_minimiser_position, const size_t right_minimiser_position)
 {
-    size_t span;
+    uint64_t span;
     if constexpr (level == 1)
         span = span1;
     if constexpr (level == 2)
@@ -592,7 +592,7 @@ inline bool RSHash3::check(const uint64_t kmer, const uint64_t kmer_rc,
 }
 
 
-inline bool RSHash3::extend_in_text(size_t &text_pos, size_t start, size_t end,
+inline bool RSHash3::extend_in_text(uint64_t &text_pos, uint64_t start, uint64_t end,
     bool forward, const uint64_t query, const uint64_t query_rc)
 {
     if(forward) {
@@ -614,7 +614,7 @@ inline bool RSHash3::extend_in_text(size_t &text_pos, size_t start, size_t end,
 
 
 template<int level>
-inline void RSHash3::fill_buffer(uint64_t *offsets, uint64_t *buffer, size_t p, size_t N, const uint64_t mask, const uint64_t shift)
+inline void RSHash3::fill_buffer(uint64_t *offsets, uint64_t *buffer, uint64_t p, uint64_t N, const uint64_t mask, const uint64_t shift)
 {
     uint64_t span;
     if constexpr (level == 1)
@@ -651,9 +651,9 @@ inline void RSHash3::fill_buffer(uint64_t *offsets, uint64_t *buffer, size_t p, 
 
 
 template<int level>
-inline bool RSHash3::check_overlap(uint64_t skmer_pos, uint64_t text_pos, size_t& start_pos, size_t& end_pos)
+inline bool RSHash3::check_overlap(uint64_t skmer_pos, uint64_t text_pos, uint64_t& start_pos, uint64_t& end_pos)
 {
-    size_t span;
+    uint64_t span;
     if constexpr (level == 1)
         span = span1;
     if constexpr (level == 2)
@@ -672,7 +672,7 @@ template<int level>
 inline bool RSHash3::check_minimiser_pos(uint64_t *buffer, const uint64_t offset,
     const uint64_t query, const uint64_t queryrc,
     const size_t s, const size_t e, const size_t minimiser_pos,
-    bool &forward, size_t &text_pos, size_t &start_pos, size_t &end_pos)
+    bool &forward, uint64_t &text_pos, uint64_t &start_pos, uint64_t &end_pos)
 {
     if(buffer[s+minimiser_pos] == queryrc) {
         forward = false;
@@ -695,7 +695,7 @@ template<int level>
 inline bool RSHash3::check_minimiser_pos2(uint64_t *buffer, const uint64_t offset,
     const uint64_t query, const uint64_t queryrc,
     const size_t s, const size_t e, const size_t left_minimiser_pos, const size_t right_minimiser_pos,
-    bool &forward, size_t &text_pos, size_t &start_pos, size_t &end_pos)
+    bool &forward, uint64_t &text_pos, uint64_t &start_pos, uint64_t &end_pos)
 {
     if(buffer[s+left_minimiser_pos] == queryrc) {
         forward = false;
@@ -729,8 +729,8 @@ inline bool RSHash3::check_minimiser_pos2(uint64_t *buffer, const uint64_t offse
 template<int level>
 inline bool RSHash3::lookup_buffer(uint64_t* buffer, uint64_t *offsets, const size_t no_skmers,
     const uint64_t query, const uint64_t queryrc,
-    size_t &text_pos, const size_t left_minimiser_pos, const size_t right_minimiser_pos,
-    bool &forward, size_t &start_pos, size_t &end_pos)
+    uint64_t &text_pos, const size_t left_minimiser_pos, const size_t right_minimiser_pos,
+    bool &forward, uint64_t &start_pos, uint64_t &end_pos)
 {
     size_t span, m;
     if constexpr (level == 1) {
@@ -870,9 +870,9 @@ uint64_t RSHash3::streaming_query(const seqan3::bitpacked_sequence<seqan3::dna4>
     uint64_t* buffer1 = new uint64_t[(m_thres1-1) * span1];
     uint64_t* buffer2 = new uint64_t[(m_thres2-1) * span2];
     uint64_t* buffer3 = new uint64_t[(m_thres3-1) * span3];
-    size_t no_skmers1, no_skmers2, no_skmers3;
-    size_t unitig_begin, unitig_end;
-    size_t text_pos;
+    uint64_t no_skmers1, no_skmers2, no_skmers3;
+    uint64_t unitig_begin, unitig_end;
+    uint64_t text_pos;
     bool forward;
     bool found = false;
     bool rolling1 = false;
